UPSTREAM: 10707: logs: Use resource builder
